+++ include/lpj.h
@@ -9,7 +9,7 @@
 /**     The model simulates vegetation dynamics, hydrology and soil                \n**/
 /**     organic matter dynamics on an area-averaged grid cell basis using          \n**/
 /**     one-year time step. Input parameters are monthly mean air                  \n**/
-/**     temperature, total precipitation and percentage of full sunshine,          \n**/
+/**     temperature, total precipitation, short and longwave radiation,            \n**/
 /**     annual atmospheric CO2 concentration and soil texture class. The           \n**/
 /**     simulation for each grid cell begins from "bare ground",                   \n**/
 /**     requiring a "spin up" (under non-transient climate) of c. 1000             \n**/