+++ src/climate/interpolate_climate.c
@@ -23,10 +23,6 @@ void interpolate_climate(Climate *climate, int index, Real x)
   printf("in=%g %g %g %g\n", climate->data[0].wet[0], climate->data[index + 1].wet[0], climate->data[(index + 1) % 2 + 1].wet[0], x);
 #endif
   interpolate_data(climate->data[0].prec, climate->data[index + 1].prec, climate->data[(index + 1) % 2 + 1].prec, climate->file_prec.n, x);
-  if (climate->data[0].sun != NULL)
-    interpolate_data(climate->data[0].sun, climate->data[index + 1].sun, climate->data[(index + 1) % 2 + 1].sun, climate->file_cloud.n, x);
-  if (climate->data[0].lwnet != NULL)
-    interpolate_data(climate->data[0].lwnet, climate->data[index + 1].lwnet, climate->data[(index + 1) % 2 + 1].lwnet, climate->file_lwnet.n, x);
-  if (climate->data[0].swdown != NULL)
-    interpolate_data(climate->data[0].swdown, climate->data[index + 1].swdown, climate->data[(index + 1) % 2 + 1].swdown, climate->file_swdown.n, x);
+  interpolate_data(climate->data[0].lwnet, climate->data[index + 1].lwnet, climate->data[(index + 1) % 2 + 1].lwnet, climate->file_lwnet.n, x);
+  interpolate_data(climate->data[0].swdown, climate->data[index + 1].swdown, climate->data[(index + 1) % 2 + 1].swdown, climate->file_swdown.n, x);
 } /* of 'interpolate_climate' */