+++ src/lpj/fscanconfig.c
@@ -207,7 +207,7 @@ Bool fscanconfig(Config *config,    /**< LPJ configuration */
   char *landuse[]={"no","yes","const","all_crops","only_crops"};
   char *fertilizer[]={"no","yes","auto"};
   char *irrigation[]={"no","lim","pot","all"};
-  char *radiation[]={"cloudiness","radiation","radiation_swonly","radiation_lwdown"};
+  char *radiation[]={"radiation","radiation_lwdown"};
   char *fire[]={"no_fire","fire","spitfire","spitfire_tmax"};
   char *sowing_data_option[]={"no_fixed_sdate","fixed_sdate","prescribed_sdate","prescribed_all_rainfed_sdate","prescribed_all_irrig_sdate"};
   char *soilpar_option[]={"no_fixed_soilpar","fixed_soilpar","prescribed_soilpar"};
@@ -300,7 +300,7 @@ Bool fscanconfig(Config *config,    /**< LPJ configuration */
 #endif
   if(fscankeywords(file,&config->unlim_nitrogen,"with_nitrogen",nitrogen,2,!config->pedantic,verbose))
     return TRUE;
-  if(fscankeywords(file,&config->with_radiation,"radiation",radiation,4,FALSE,verbose))
+  if(fscankeywords(file,&config->radiation_lwdown,"radiation",radiation,2,FALSE,verbose))
     return TRUE;
 #if defined IMAGE && defined COUPLED
   if(config->sim_id==LPJML_IMAGE && config->with_radiation)
@@ -365,11 +365,10 @@ Bool fscanconfig(Config *config,    /**< LPJ configuration */
     if(fscanbool(file,&config->with_glaciers,"with_glaciers",!config->pedantic,verbose))
       return TRUE;
   }
-  if(config->isanomaly && config->with_radiation!=RADIATION)
+  if(config->isanomaly && config->radiation_lwdown)
   {
     if(isroot(*config))
-      fprintf(stderr,"ERROR208: Radiation setting '%s' not supported for anomalies, must be 'radiation'.\n",
-              radiation[config->with_radiation]);
+      fprintf(stderr,"ERROR208: Radiation setting 'radiation_lwdown' not supported for anomalies, must be 'radiation'.\n");
     return TRUE;
   }
   config->percolation_heattransfer = TRUE;
@@ -893,46 +892,19 @@ Bool fscanconfig(Config *config,    /**< LPJ configuration */
     }
   }
   scanclimatefilename(input,&config->temp_filename,TRUE,TRUE,"temp");
+  scanclimatefilename(input,&config->prec_filename,TRUE,TRUE,"prec");
+  scanclimatefilename(input,&config->lwnet_filename,TRUE,TRUE,config->radiation_lwdown ? "lwdown" : "lwnet");
+  scanclimatefilename(input,&config->swdown_filename,TRUE,TRUE,"swdown");
   if (config->isanomaly)
   {
+    scanclimatefilename(input, &config->delta_lwnet_filename,FALSE,FALSE,config->radiation_lwdown ? "delta_lwdown" : "delta_lwnet");
+    scanclimatefilename(input, &config->delta_swdown_filename,FALSE,FALSE, "delta_swdown");
     scanclimatefilename(input, &config->delta_temp_filename,FALSE,TRUE,"delta_temp");
     if(config->with_glaciers)
     {
       scanclimatefilename(input,&config->icefrac_filename,TRUE,TRUE,"icefrac");
     }
     scanclimatefilename(input, &config->delta_prec_filename,FALSE,TRUE, "delta_prec");
-  }
-  scanclimatefilename(input,&config->prec_filename,TRUE,TRUE,"prec");
-  switch(config->with_radiation)
-  {
-    case RADIATION:
-      scanclimatefilename(input,&config->lwnet_filename,TRUE,TRUE,"lwnet");
-      if (config->isanomaly)
-        scanclimatefilename(input, &config->delta_lwnet_filename,FALSE,FALSE, "delta_lwnet");
-     scanclimatefilename(input,&config->swdown_filename,TRUE,TRUE,"swdown");
-     if (config->isanomaly)
-       scanclimatefilename(input, &config->delta_swdown_filename,FALSE,FALSE, "delta_swdown");
-      break;
-    case RADIATION_LWDOWN:
-      scanclimatefilename(input,&config->lwnet_filename,TRUE,TRUE,"lwdown");
-      if (config->isanomaly)
-        scanclimatefilename(input, &config->delta_lwnet_filename,FALSE,FALSE, "delta_lwdown");
-      scanclimatefilename(input,&config->swdown_filename,TRUE,TRUE,"swdown");
-      if (config->isanomaly)
-        scanclimatefilename(input, &config->delta_swdown_filename,FALSE,FALSE, "delta_swdown");
-      break;
-    case CLOUDINESS:
-      scanclimatefilename(input,&config->cloud_filename,TRUE,TRUE,"cloud");
-      break;
-    case RADIATION_SWONLY:
-      scanclimatefilename(input,&config->swdown_filename,TRUE,TRUE,"swdown");
-      if (config->isanomaly)
-        scanclimatefilename(input, &config->delta_swdown_filename,FALSE,TRUE, "delta_swdown");
-     break;
-    default:
-      if(verbose)
-        fprintf(stderr,"ERROR213: Invalid setting %d for radiation.\n",config->with_radiation);
-      return TRUE;
   }
   if(!config->unlim_nitrogen && !config->no_ndeposition)
   {